requirements.txt

flask==3.0.3
requests==2.32.3
numpy==1.26.4
pandas==2.2.2
streamlit==1.38.0
python-dotenv==1.0.1
bs4==0.0.1
selenium==3.141.0
webdriver-manager==4.0.2
google-generativeai==0.8.4